Create README.md
....Casino Number Guessing Game...
The program asks for a betting amount and then asks the user to guess a number on rolling. If the random number generated matches the user input, he wins, else money is deducted. The user can keep playing until he loses all the amount he put in initially. 
Here in the programming language we have a random number instead of Guessing the number in real life. Let explanation our code step by step, First, we will take a Username and the total amount of Money in the Deposit, Now the player can play a Casino Game or Number Guessing Game but before playing a game player need to Bid for every time.
Explanation:
Suppose I am a Player My Name is Ghanendra and I have a total of 5000$ Money. I deposited My total money and Strat playing a Casino Game or Number Guessing Game First I took a Bid 1000$ and I guessed the Number 7(Between 1 to 10) if the random number generated by the computer is 7 then I can win 10 times of My Bid amount and if I lose My remaining Money is 4000$. Now I can play more until I have a single penny or I can Win or Maybe not.